+++ src/net/network_manager.cpp
@@ -4,6 +4,7 @@
 #include <managers/error_queues.hpp>
 #include <managers/account.hpp>
 #include <managers/profile_cache.hpp>
+#include <managers/friend_list.hpp>
 #include <util/net.hpp>
 #include <util/debug.hpp>
 
@@ -50,6 +51,12 @@ NetworkManager::NetworkManager() {
         log::info("Successfully logged into the server!");
         connectedTps = packet->tps;
         _loggedin = true;
+
+        // additionally, preload the friendlist
+        auto& flm = FriendListManager::get();
+        if (!flm.isLoaded()) {
+            flm.load();
+        }
     });
 
     addBuiltinListener<LoginFailedPacket>([this](auto packet) {