+++ src/util/ui.hpp
@@ -18,4 +18,7 @@ namespace util::ui {
 
     // set the scale for `node` in a way that `node->getScaledContentSize()` will match `targetSize`
     void rescaleToMatch(cocos2d::CCNode *node, cocos2d::CCSize targetSize, bool stretch = false);
+
+    float getScrollPos(BoomListView* listView);
+    void setScrollPos(BoomListView* listView, float pos);
 }